requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
httpie==3.2.1
idna==3.4
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
Pillow==9.4.0
pycodestyle==2.10.0
Pygments==2.14.0
PyJWT==1.7.1
PySocks==1.7.1
pytz==2023.2
requests==2.28.2
requests-toolbelt==0.10.1
rest-framework-simplejwt==0.0.2
rich==13.3.3
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.2
urllib3==1.26.15
